complete kafka setup for Subscription
When Subscription send message for updating min and max salary, the applicant service will consume and store new updated min max salary to Resume collection with corresponding resumeId

+++ backend/subscription/src/main/java/com/devision/subscription/kafka/ApplicantProfileUpdateProducer.java
@@ -5,6 +5,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Component
@@ -20,7 +21,7 @@ public class ApplicantProfileUpdateProducer {
         this.topic = topic;
     }
 
-    public void publishSalaryUpdate(String applicantId, Integer minSalary, Integer maxSalary) {
+    public void publishSalaryUpdate(String applicantId, BigDecimal minSalary, BigDecimal maxSalary) {
         ApplicantProfileUpdateEvent event = new ApplicantProfileUpdateEvent(
                 applicantId,
                 minSalary,
